+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add return typehints to the methods of the `SentryExtension` class to prepare for Symfony 6 (#563)
 - Fix setting the IP address on the user context when it's not available (#565)
+- Fix wrong method existence check in `TracingDriverConnection::errorCode()` (#568)
 
 ## 4.2.3 (2021-09-21)